overflow

+++ src/main/java/scorecollection/CreditHistory.java
@@ -18,12 +18,20 @@ public class CreditHistory {
       ratings.add(new CreditRating(rating, LocalDate.now()));
    }
 
+   // START:overflow
    public int arithmeticMean() {
       if (ratings.size() == 0) throw new IllegalStateException();
 
-      var total = ratings.stream().mapToInt(CreditRating::rating).sum();
-      return total / ratings.size();
+      var total = ratings.stream()
+      // START_HIGHLIGHT
+                     .mapToLong(CreditRating::rating)
+      // END_HIGHLIGHT
+                     .sum();
+      // START_HIGHLIGHT
+      return (int)(total / ratings.size());
+      // END_HIGHLIGHT
    }
+   // END:overflow
 
    public long daysSpanned() {
       if (ratings.isEmpty()) return 0;
